Module mezzaluna_feature_loader::loaded_features

A set to track loaded Ruby source paths based on a Vec and HashSet.

Ruby tracks which files and native extensions have been required in a global variable called $LOADED_FEATURES, which is aliased to $". $LOADED_FEATURES is an Array of paths which point to these Ruby sources and native extensions.

This module exposes an append-only, insertion order-preserving, set-like container for tracking disk and in-memory Ruby sources as they are evaluated on a Ruby interpreter using require and require_relative.

See LoadedFeatures for more documentation on how to use the types in this module.

Structs

Features

An iterator over the features in a LoadedFeatures.

Iter

An iterator over the feature paths in a LoadedFeatures.

LoadedFeatures

A set of all sources loaded by a Ruby interpreter with require and require_relative.