Crate mezzaluna_feature_loader

Ruby feature loader.

An Artichoke Ruby VM may load code (called “features”) with several strategies. Features can be loaded from an in-memory virtual file system which can also store native extensions, natively from local disk, or via a set of search paths given by the RUBYLIB environment variable on interpreter boot.

Modules

loaded_features

A set to track loaded Ruby source paths based on a Vec and HashSet.

loaders

Strategies for loading Ruby source code and native extensions.

paths

Path routines for interfacing with load paths.

Structs

Feature

A Ruby source (“feature”) that has been loaded into an interpreter.

Handledisk

A handle to a file that can be tested for equality with other handles.

LoadedFeatures

A set of all sources loaded by a Ruby interpreter with require and require_relative.