Struct mezzaluna_feature_loader::loaders::Disk

pub struct Disk { /* private fields */ }

Available on crate feature disk only.

A Ruby source code loader that loads sources directly from disk and resolves relative paths with the Ruby $LOAD_PATH.

MRI Ruby allows manipulating the require search path by modifying the $LOAD_PATH global, or its alias $:, at runtime.

MRI Ruby allows requiring sources with relative, explicit relative or absolute paths.

Relative paths are paths of the form json/pure which do not begin with either a file system root like / or C:\ or an explicit relative directory marker like .. or .. These relative paths are resolved relative to the Ruby load path.

$LOAD_PATH contains a list of directory paths to search for Ruby sources and behaves similarly to the RUBYLIB environment variable in the Rubylib loader.

// Search `/home/artichoke/src` first, only attempting to search
// `/usr/share/artichoke` if no file is found in `/home/artichoke/src`.
//
// The relative path `./_lib` is resolved relative to the given working
// directory.
let fixed_loader = Disk::with_load_path_and_cwd(
    vec![PathBuf::from("/home/artichoke/src"), PathBuf::from("/usr/share/artichoke"), PathBuf::from("_lib")],
    Path::new("/home/artichoke")
)?;

Implementations

impl Disk

pub const fn new() -> Self

Create a new native file system loader that searches the file system for Ruby sources with an empty $LOAD_PATH.

A Disk loader with an empty $LOAD_PATH can only load sources by absolute paths or relative to the process’s current working directory if an explicit relative path is given.

The resolved load paths are mutable; $LOAD_PATH can be modified at runtime by Ruby code as the VM executes. See load_path and set_load_path for reading and modifying a Disk loader’s load path.

This source loader grants access to the host file system. The Disk loader does not support native extensions.

pub fn with_load_path(load_path: Vec<PathBuf>) -> Option<Self>

Create a new native file system loader that searches the file system for Ruby sources at the paths specified by the given $LOAD_PATH from the Ruby global variable.

The resolved load paths are mutable; $LOAD_PATH can be modified at runtime by Ruby code as the VM executes. See load_path and set_load_path for reading and modifying a Disk loader’s load path.

If any of the paths in the $LOAD_PATH global variable are not absolute paths, they are absolutized relative to the current process’s current working directory at the time the load path is set or modified.

This source loader grants access to the host file system. The Disk loader does not support native extensions.

This method returns None if the current working directory cannot be retrieved, or if the $LOAD_PATH global variable does not contain any paths.

pub fn with_load_path_and_cwd(
load_path: Vec<PathBuf>,
cwd: &Path
) -> Option<Self>

Create a new native file system loader that searches the file system for Ruby sources at the paths specified by the given load_path platform string.

The resolved load paths are mutable; $LOAD_PATH can be modified at runtime by Ruby code as the VM executes. See load_path and set_load_path for reading and modifying a Disk loader’s load path.

If any of the paths in the $LOAD_PATH global variable are not absolute paths, they are absolutized relative to the current process’s current working directory at the time the load path is set or modified.

This source loader grants access to the host file system. The Disk loader does not support native extensions.

This method returns None if the given load_path does not contain any paths.

pub fn resolve_file(&self, path: &Path) -> Option<Handle>

Check whether path points to a file in the backing file system and return a file Handle if it exists.

Returns Some if the file system object pointed to by path exists. If path is relative, it is joined to each path in the $LOAD_PATH environment variable at the time this loader was initialized. If path is an explicit relative path, it is joined with the current working directory.

This method is infallible and will return None for non-existent paths or if the current working directory cannot be resolved when given an explicit relative path.

pub fn load_path(&self) -> &[PathBuf]

Return a reference to the loader’s current $LOAD_PATH.

Examples

let loader = Disk::with_load_path_and_cwd(
    vec![PathBuf::from("/home/artichoke/src"), PathBuf::from("/usr/share/artichoke"), PathBuf::from("_lib")],
    Path::new("/home/artichoke"),
)?;
assert_eq!(
    loader.load_path(),
    &[Path::new("/home/artichoke/src"), Path::new("/usr/share/artichoke"), Path::new("/home/artichoke/_lib")]
);

pub fn set_load_path(
&mut self,
load_path: Vec<PathBuf>,
cwd: &Path
) -> Vec<PathBuf>

Replace the loader’s current $LOAD_PATH.

This method returns the loader’s previous $LOAD_PATH.

Examples

let mut loader = Disk::with_load_path_and_cwd(
    vec![PathBuf::from("/home/artichoke/src"), PathBuf::from("/usr/share/artichoke"), PathBuf::from("_lib")],
    Path::new("/home/artichoke"),
)?;
assert_eq!(
    loader.load_path(),
    &[Path::new("/home/artichoke/src"), Path::new("/usr/share/artichoke"), Path::new("/home/artichoke/_lib")]
);

let old_load_path = loader.set_load_path(
    vec![PathBuf::from("libpath")],
    Path::new("/home/app"),
);
assert_eq!(
    old_load_path,
    &[Path::new("/home/artichoke/src"), Path::new("/usr/share/artichoke"), Path::new("/home/artichoke/_lib")]
);
assert_eq!(loader.load_path(), [Path::new("/home/app/libpath")]);